Issue 45827: Sample Finder saved searches are not showing up on trial
