+++ api/src/org/labkey/api/study/StudyService.java
@@ -29,6 +29,7 @@ import org.labkey.api.module.Module;
 import org.labkey.api.query.QuerySchema;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.reports.model.ViewCategory;
+import org.labkey.api.reports.report.view.ReportUtil;
 import org.labkey.api.security.SecurableResource;
 import org.labkey.api.security.User;
 import org.labkey.api.services.ServiceRegistry;
@@ -235,4 +236,6 @@ public interface StudyService
     Collection<String> getParticipantIds(Study study, User user);
 
     boolean participantExists(Study study, String participantId);
+
+    ReportUtil.ReportFilter getStudyReportFilter(boolean editOnly);
 }
